Refactor EvoEngine to explicit application-owned runtime state with install deployment, runtime package support, refreshed docs, LF formatting, and SDK warning cleanup.
refactor: move runtime ownership into Application
feat: add runtime package loading infrastructure
build: add install deployment and simplify VS presets
docs: refresh README and plugin documentation
chore: add LF normalization and C++ formatting helper
fix: clean SDK narrowing warnings

+++ EvoEngine_Plugins/EcoSysLab/include/MeshFormation/TreeSkinnedMeshGenerator.hpp
@@ -89,7 +89,7 @@ void CylindricalSkinnedMeshGenerator<SkeletonData, FlowData, NodeData>::Generate
   std::vector<std::vector<std::pair<SkeletonNodeHandle, int>>> temp_steps{};
   temp_steps.resize(Jobs::GetWorkerSize());
 
-  Jobs::RunParallelFor(sorted_internode_list.size(), [&](unsigned internode_index, unsigned thread_index) {
+  Jobs::RunParallelFor(sorted_internode_list.size(), [&](size_t internode_index, size_t thread_index) {
     auto internode_handle = sorted_internode_list[internode_index];
     const auto& internode = skeleton.PeekNode(internode_handle);
     const auto& internode_info = internode.info;
